+++ CHANGELOG.md
@@ -6,10 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 2.2.0 - 2025-07-31
+## 2.2.0 - 2025-10-16
+
+### Fixed
+- Editor initialization failure handling. #INT-3365
 
 ### Changed
-- Set the default cloudChannel to 8
+- Set the default cloudChannel to 8. #INT-3357
 
 ## 2.1.0 - 2023-03-27